Add files via upload
This challenge is near to my heart. Picture this: It's the mid-90s. I'm a teenager and my mother just bought us a Packard-Bell and I am just enamored. It ran Windows 3.1 and was already an outdated model, but I didn't care (or know the difference, to be honest.) I'm trolling fools on Compu-Serve, I'm reading the newsgroups, I'm making crude pictures in Corel Draw. I get into learning everything I possibly can about the world because now it's right there for the taking. I start reading Phrack, and one day there this article in #49 written by Aleph One titled, "Smashing The Stack For Fun And Profit"

Now, I was in Junior High School and I was just taking my very first BASIC programming class. I had accidentally set a BIOS password on the Packard-Bell and I had to read the manual to pull the jumper to wipe it. That was the full extent of my experience with these subjects. But when I read that article, it gave me a hint of another world. Well, I just had to try it.

I'm sure if you're reading this then you know all about this topic, and you've probably already done this exploit many times, but I'm going to include this challenge for nostalgia's sake. Here we have a faithful reproduction of the code I wrote over the week after I read the article. It took me a long time because I had to teach myself C syntax and compiling to do it. It's not much, but I've always had a soft spot for this.

Here's your target: 35.184.182.18 32003 and the target address for the running binary is 0x401162.